handing submition to the backend in RuleEditor
