+++ public/pages/Rules/containers/CreateRule/CreateRule.tsx
@@ -6,14 +6,11 @@
 import { BrowserServices } from '../../../../models/interfaces';
 import { RuleEditor } from '../../components/RuleEditor/RuleEditor';
 import React, { useContext } from 'react';
-import { EuiButton, EuiFlexGroup, EuiFlexItem } from '@elastic/eui';
 import { RouteComponentProps } from 'react-router-dom';
-import { BREADCRUMBS, ROUTES } from '../../../../utils/constants';
-import { Rule } from '../../../../../models/interfaces';
+import { BREADCRUMBS } from '../../../../utils/constants';
 import { CoreServicesContext } from '../../../../components/core_services';
 import { NotificationsStart } from 'opensearch-dashboards/public';
-import { errorNotificationToast } from '../../../../utils/helpers';
-import { setBreadCrumb, validateRule } from '../../utils/helpers';
+import { setBreadCrumb } from '../../utils/helpers';
 
 export interface CreateRuleProps {
   services: BrowserServices;
@@ -24,40 +21,14 @@ export interface CreateRuleProps {
 export const CreateRule: React.FC<CreateRuleProps> = ({ history, services, notifications }) => {
   const context = useContext(CoreServicesContext);
   setBreadCrumb(BREADCRUMBS.RULES_CREATE, context?.chrome.setBreadcrumbs);
-  const footerActions: React.FC<{ rule: Rule }> = ({ rule }) => {
-    const onCreate = async () => {
-      if (!validateRule(rule, notifications!, 'create')) {
-        return;
-      }
-      const createRuleRes = await services.ruleService.createRule(rule);
-
-      if (!createRuleRes.ok) {
-        errorNotificationToast(notifications!, 'create', 'rule', createRuleRes.error);
-      } else {
-        history.replace(ROUTES.RULES);
-      }
-    };
-
-    return (
-      <EuiFlexGroup justifyContent="flexEnd">
-        <EuiFlexItem grow={false}>
-          <EuiButton onClick={() => history.replace(ROUTES.RULES)}>Cancel</EuiButton>
-        </EuiFlexItem>
-        <EuiFlexItem grow={false}>
-          <EuiButton fill onClick={onCreate} data-test-subj={'create_rule_button'}>
-            Create
-          </EuiButton>
-        </EuiFlexItem>
-      </EuiFlexGroup>
-    );
-  };
 
   return (
     <RuleEditor
       title="Create a rule"
       history={history}
       notifications={notifications}
-      FooterActions={footerActions}
+      mode={'create'}
+      ruleService={services.ruleService}
     />
   );
 };
