+++ public/pages/Rules/utils/constants.ts
@@ -24,4 +24,4 @@ export const ruleSeverity: { name: string; value: string }[] = [
 
 export const ruleSource: string[] = ['Sigma', 'Custom'];
 
-export const ruleStatus: string[] = ['Select a rule status', 'experimental', 'test', 'stable'];
+export const ruleStatus: string[] = ['experimental', 'test', 'stable'];
